StorageComponent door toggling works.
